comentarios e login
add os comentarios pros usuarios login p usuario e p admin e tals

+++ controllers/postController.js
@@ -1,4 +1,7 @@
 const Post = require("../models/posts");
+const Comentario = require("../models/comentarios");
+
+
 
 
 // mostrar todos os posts
@@ -13,18 +16,28 @@ exports.listarPosts = async (req, res) => {
             .sort({ data: -1 });
 
 
+
         console.log("posts encontrados:", posts.length);
 
 
+
         res.render("posts", {
+
             posts: posts
+
         });
+
 
 
     } catch (err) {
 
+
         console.log(err);
-        res.status(500).send("Erro ao carregar posts");
+
+
+        res.status(500)
+            .send("Erro ao carregar posts");
+
 
     }
 
@@ -34,10 +47,24 @@ exports.listarPosts = async (req, res) => {
 
 
 
-// criar novo post
+
+
+
+
+// criar novo post (ADMIN)
 exports.criarPost = async (req, res) => {
 
     try {
+
+
+        if (!req.session.admin) {
+
+            return res.redirect("/login");
+
+        }
+
+
+
 
         const novoPost = new Post({
 
@@ -50,16 +77,26 @@ exports.criarPost = async (req, res) => {
         });
 
 
+
+
         await novoPost.save();
+
+
 
 
         res.redirect("/posts");
 
 
+
+
     } catch (err) {
 
+
         console.log(err);
-        res.status(500).send("Erro ao criar post");
+
+
+        res.status(500)
+            .send("Erro ao criar post");
 
     }
 
@@ -67,20 +104,130 @@ exports.criarPost = async (req, res) => {
 
 
 
-// apagar post
+
+
+
+
+
+
+// apagar post (ADMIN)
 exports.apagarPost = async (req, res) => {
 
     try {
 
+
+        if (!req.session.admin) {
+
+            return res.redirect("/login");
+
+        }
+
+
+
+
         await Post.findByIdAndDelete(req.params.id);
+
+
+
 
         res.redirect("/posts");
 
 
+
+
     } catch (err) {
 
+
         console.log(err);
-        res.status(500).send("Erro ao apagar");
+
+
+        res.status(500)
+            .send("Erro ao apagar");
+
+
+    }
+
+};
+
+
+
+
+
+
+
+
+
+
+
+// mostrar um post específico
+exports.verPost = async (req, res) => {
+
+    try {
+
+
+
+        const post = await Post.findById(req.params.id);
+
+
+
+
+        if (!post) {
+
+
+            return res.status(404)
+                .send("Post não encontrado");
+
+
+        }
+
+
+
+
+
+
+        const comentarios = await Comentario.find({
+
+            postId: post._id
+
+        })
+        .sort({
+
+            data: -1
+
+        });
+
+
+
+
+
+
+
+
+        res.render("post", {
+
+
+            post: post,
+
+            comentarios: comentarios
+
+
+        });
+
+
+
+
+
+    } catch (err) {
+
+
+        console.log(err);
+
+
+
+        res.status(500)
+            .send("Erro ao carregar post");
+
+
 
     }
